requirements.txt
asgiref==3.4.1
autopep8==1.6.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==1.11.29
django-autocomplete-light==3.2.10
django-ckeditor==5.4.0
django-crispy-forms==1.13.0
django-formtools==2.3
django-import-export==2.6.1
django-js-asset==1.2.2
django-redis==4.9.0
django-reversion==4.0.1
django-silk==3.0.0
djangorestframework==3.8.2
et-xmlfile==1.1.0
future==0.18.2
gprof2dot==2016.10.13
hiredis==0.2.0
httplib2==0.9.2
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupPy==1.14
MarkupSafe==2.0.1
mistune==2.0.0
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.10.0
PyMySQL==1.0.2
pypiserver==1.4.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.0.2
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tablib==3.1.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
uuid==1.30
wrapt==1.13.3
xlrd==2.0.1
xlwt==1.3.0